Moved is_comment to a global namespace
